embedder/requirements/common.txt
# Scraping
requests==2.31.0
beautifulsoup4==4.12.3
validators==0.22.0

# Embedding
langchain==0.1.1
sentence-transformers==2.2.2

# Config
python-decouple==3.8

# CLI
typer==0.9.0
rich==13.7.0
sqlite-vss==0.1.2